+++ src/init.c
@@ -27,6 +27,7 @@ extern SEXP _prioritizr_rcpp_apply_min_shortfall_objective(SEXP, SEXP, SEXP, SEX
 extern SEXP _prioritizr_rcpp_apply_neighbor_constraints(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_boundary_data(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_branch_matrix(SEXP);
+extern SEXP _prioritizr_rcpp_ferrier_score(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_forbid_solution(SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_get_optimization_problem_A(SEXP);
 extern SEXP _prioritizr_rcpp_get_optimization_problem_col_ids(SEXP);
@@ -73,6 +74,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_prioritizr_rcpp_apply_neighbor_constraints",                        (DL_FUNC) &_prioritizr_rcpp_apply_neighbor_constraints,                        3},
     {"_prioritizr_rcpp_boundary_data",                                     (DL_FUNC) &_prioritizr_rcpp_boundary_data,                                     4},
     {"_prioritizr_rcpp_branch_matrix",                                     (DL_FUNC) &_prioritizr_rcpp_branch_matrix,                                     1},
+    {"_prioritizr_rcpp_ferrier_score",                                     (DL_FUNC) &_prioritizr_rcpp_ferrier_score,                                     3},
     {"_prioritizr_rcpp_forbid_solution",                                   (DL_FUNC) &_prioritizr_rcpp_forbid_solution,                                   2},
     {"_prioritizr_rcpp_get_optimization_problem_A",                        (DL_FUNC) &_prioritizr_rcpp_get_optimization_problem_A,                        1},
     {"_prioritizr_rcpp_get_optimization_problem_col_ids",                  (DL_FUNC) &_prioritizr_rcpp_get_optimization_problem_col_ids,                  1},